Unittests Completed

+++ app.py
@@ -281,7 +281,7 @@ def add_like(message_id):
     """Add a like on the message"""
 
     if not g.user:
-        flash("Login to like a Warbler post.", "danger")
+        flash("Access unauthorized.", "danger")
         return redirect("/")
 
     like = Likes(user_id=g.user.id, message_id=message_id)
@@ -362,8 +362,8 @@ def messages_add():
 @app.route('/messages/<int:message_id>', methods=["GET"])
 def messages_show(message_id):
     """Show a message."""
-
-    msg = Message.query.get(message_id)
+    
+    msg = Message.query.get_or_404(message_id)
     return render_template('messages/show.html', message=msg)
 
 
@@ -377,6 +377,11 @@ def messages_destroy(message_id):
         return redirect("/")
 
     msg = Message.query.get(message_id)
+
+    if msg.user_id != g.user.id:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
+
     db.session.delete(msg)
     db.session.commit()
